Simplified color handling (v5.0.0) (#103)
Following the recent firmware update, the gateway now returns more information around colors. This is an attempt to simplify returning what is supported by a bulb. I'm having to go for a breaking change as I need to move methods around in between classes. The color values that are returned seem to be hue, saturation, x and y.

If the user wants to work in the RGB color space, it will have to do the conversion itself. Hence the color example file.

Change log:

color.py:
- Added hex to rgb converter
- Added method that returns what features are supported by a bulb (dimmer, hex, xy, mireds)
- Removed abundant methods

const.py
- Found a set of new constants and added them
- ATTR_LIGHT_COLOR (5706) is hex so I renamed the constant
- Reshuffled constants from other files

device.py
- Removed inferred methods
- Removed RGB color conversion alltogether
- Only return attributes that are relevant for the bulb (ie a white spectrum bulb does return a color temperature, not a rgb value)
- Reordered setters and getters so that they are in the same order in the different classes

Tests
- Moved the test bulbs into a separate file and also added motion sensor and dimmer button devices.
Other:
- Updated tests