Fixes and optimizes the Docker build
This resolves issue #203 and also optimizes the Dockerfile to use less
resources. Also the current version from the Git repo is copied to
the Docker container for easy access.

I've taken the liberty to also update the DTLSSocket requirement to
version 0.1.7 which resolves an issue regarding cython. According to
their repo, nothing else has changed [1].

The RPi and RPiW Dockerfiles were not updated because of compatibility
reasons. I cannot easily test those files and do not want to break
anything.

**References**

 1. https://git.fslab.de/jkonra2m/tinydtls-cython/commits/master