A bit more stable observations
This change fixes a very small bug (I suppose) in the `aiocoap_api` class
where the flag whether to parse the output as json was not there in the
observe method. Hopefully this makes this part of the API a bit more stable.

Also I've added an observation callback as a global variable in the
`ApiResource` class. When tracking this value for an object, it is possible
to invoke the callback manually. This way the observation can also be
invoked when the user makes a manual change. Thus don't having the need to
call out the observation.

Thirdly I've added an explicit error callback to the `ApiResource` observation
which can be used to diagnose issues easily with the use of the print function.

Finally I've also added a check to check whether or not the incoming result has
even changed from the already existing result. Thus eliminating the need to
continiously update the value. This can hopefully make some parts of the
library a little more stable, because it will limit callback calls.

In conclusion, these are very minor changes, but hopefully will make the
library and in turn the applications using this library a little bit more
stable.