Added cancellation of obervations
This adds a (non-breaking) cancel command to the API. This makes it possible to
cancel individual observations easily by sending a `observe_cancel` command to
the `aiocoap` API.

I have included a rudimentary example in the
[example_async.py](examples/example_async.py) file.

This also fixes a very small 'bug' which was triggered when the API was
shutdown. The logging about incorrectly ending the loop is now gone.