Resets an observation when a CHANGED is received
It seems that the `pytradfri` lib tends to get stuck on observations for
some people. This change will throw an Error when a CHANGED message is
received. According to some users they only receive these messages when
the observation is in a stuck state.

When implemented correctly - which it is on `example_async.py` - the
observation will restart. I tried this myself by forcing a reset of an
observation after 5 messages and that seems to work beautifully.

Feedback is appreciated!