Observation reset (#214)
* Observation reset

This commit resets the observations when they hit a timeout, and also prevents
duplicate state changes. Also observations are cancelled whenever possible
instead of erroring out.

 1. Whenever the light or socket already contains a specific state (ie. the
    light is on and is dimmed at 20%), and a change is requested, a filter will
    be applied and only the changed states will be send to the device.
 2. Observations are reset whenever no state change was received for a limited
    amount of time. This is dependent on the first change, because Tradfri does
    not send a successful callback back to us, when no change was applied.
 3. Cancelling observations instead of erroring out will create less error
    logging. There is still some error logging left, but this is because of the
    `aiocoap` library.

There is still something odd happening with observation cancellation which must
be errored out instead of cancelled, because otherwise no new observation will
be launched by Home Assistant. This is because the `aiocoap` library does not
report when a cancellation happened, but it does on errors.

* Slightly modified way of controlling the API

This slightly modifies the way a command is send to the backend API. I think
this is a more concise way, because now the `set_values` function in
`command.py` always returns a value. This works the same as the other
solution.

This also fixes the unit tests.

* Fixed light transitions

Appearently light transitions are always preceded with an on or off command.
This change removes the filter for the state attribute.

I don't know if this should be an issue in Home Assistant...

* Review fixes for #214

Some initial, small changes that came up during the review of #214. This
changes the following:

 * Better naming of checking observations
 * Cancelling observations instead of erroring them out during cancellation