aiocoap_api: prevent race in _reset_protocol (#262)
* aiocoap_api: prevent race in _reset_protocol

When multiple concurrent requests fail at the same time, (for
example when the gateway is unavailable), they will try to reset
the underlying aiocoap Protocol.  This reset operation is not
async-safe and a race condition may be triggered.

This commit marks APIFactory._reset_protocol as a critical section
and introduces a lock for preventing concurrent execution.

* aiocoap_api: remove deprecated loop argument

The loop argument to asyncio.Lock() is deprecated in python>=3.8.

* aiocoap_api: do not reset if already locked

If _reset_protocol is called more than once at the same time, we
only need to reset the protocol once.  This commit uses to lock
to detect if is the case and skips if the protocol is already being
reset.